requirements.txt

python-dotenv


Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Smorest==0.34.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.24.1